Fix slider performance: replace parseInt with parseFloat for smooth weight adjustments
- Replace parseInt(e.target.value) with parseFloat(e.target.value) in all weight sliders
- Change step from 1 to 0.1 for sub-percentage precision
- Fixes choppy 1% increment dragging, enables smooth 0.1% precision
- Affects WeightControls, PlayerDetailsModal, Players, Demo, CoachDashboard, TeamFormation
- Demo weight showcase now displays smooth professional slider movement
- Resolves slider performance issues identified in SLIDER_FIXES_SUMMARY.md

+++ frontend/src/pages/CoachDashboard.jsx
@@ -441,8 +441,8 @@ const CoachDashboard = React.memo(function CoachDashboard() {
                           min={0}
                           max={100}
                           value={percentages[drill.key]}
-                          onInput={e => updateWeightsFromPercentage(drill.key, parseInt(e.target.value))}
-                          onChange={e => updateWeightsFromPercentage(drill.key, parseInt(e.target.value))}
+                          onInput={e => updateWeightsFromPercentage(drill.key, parseFloat(e.target.value))}
+                          onChange={e => updateWeightsFromPercentage(drill.key, parseFloat(e.target.value))}
                           className="flex-1 accent-cmf-primary h-2 rounded-lg bg-gray-100"
                         />
                         <div className="w-12 text-right">
